объединил все это каким то образом. аккуратно отсюда ветвимся
