+++ app/src/main/AndroidManifest.xml
@@ -1,30 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-<<<<<<< HEAD
     package="ru.taksi.pro.android">
 
-    <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+
+<!--    <meta-data-->
+<!--        android:name="preloaded_fonts"-->
+<!--        android:resource="@array/preloaded_fonts" />-->
 
     <application
         android:name=".TaxiProApplication"
         android:allowBackup="false"
-=======
-    package="ru.ildus.taksipro">
-
-    <application
-        android:allowBackup="true"
->>>>>>> feature/view
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-<<<<<<< HEAD
         android:theme="@style/Theme.TaksiPro"
         android:usesCleartextTraffic="true">
-=======
-        android:theme="@style/AppTheme">
->>>>>>> feature/view
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -32,16 +25,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-<<<<<<< HEAD
+        <activity android:name=".ui.activitys.RegistrationActivity" />
+
     </application>
 
 </manifest>
-=======
-        <activity android:name=".registration.RegistrationActivity" />
-        <meta-data
-            android:name="preloaded_fonts"
-            android:resource="@array/preloaded_fonts" />
-    </application>
-
-</manifest>
->>>>>>> feature/view
