+++ app/src/main/java/ru/taksi/pro/android/ui/fragments/RegistrFragment.kt
@@ -1,4 +1,4 @@
-package ru.ildus.taksipro.registration
+package ru.taksi.pro.android.ui.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,19 +6,21 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.EditText
+import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import ru.ildus.taksipro.databinding.RegistrFragmentEnterCodeBinding
+import ru.taksi.pro.android.databinding.RegistrFragmentBinding
 
-class RegistrFragmentEnterCode: Fragment() {
-    private var binding: RegistrFragmentEnterCodeBinding? = null
+
+class RegistrFragment : Fragment() {
+    private var binding: RegistrFragmentBinding? = null
     private lateinit var editText: EditText
     private lateinit var button: Button
-    companion object {
-        fun newInstance() = RegistrFragmentEnterCode()
-    }
+    lateinit var context: AppCompatActivity
+
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View = RegistrFragmentEnterCodeBinding.inflate(inflater).let {
+    ): View = RegistrFragmentBinding.inflate(inflater).let {
         binding = it
         it.root
     }
@@ -26,11 +28,12 @@ class RegistrFragmentEnterCode: Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         editText = binding!!.editTextPhone
-        button = binding!!.buttonEnterCod
+        button = binding!!.buttonCod
         button.setOnClickListener{onButtonClickListener()}
     }
 
     fun onButtonClickListener() {
 
     }
+
 }
