+++ app/src/main/java/ru/taksi/pro/android/ui/fragments/RegistrFragmentEnterCode.kt
@@ -1,4 +1,4 @@
-package ru.ildus.taksipro.registration
+package ru.taksi.pro.android.ui.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,21 +6,19 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.EditText
-import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import ru.ildus.taksipro.databinding.RegistrFragmentBinding
+import ru.taksi.pro.android.databinding.RegistrFragmentEnterCodeBinding
 
-
-class RegistrFragment : Fragment() {
-    private var binding: RegistrFragmentBinding? = null
+class RegistrFragmentEnterCode: Fragment() {
+    private var binding: RegistrFragmentEnterCodeBinding? = null
     private lateinit var editText: EditText
     private lateinit var button: Button
-    lateinit var context: AppCompatActivity
-
-
+    companion object {
+        fun newInstance() = RegistrFragmentEnterCode()
+    }
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View = RegistrFragmentBinding.inflate(inflater).let {
+    ): View = RegistrFragmentEnterCodeBinding.inflate(inflater).let {
         binding = it
         it.root
     }
@@ -28,12 +26,11 @@ class RegistrFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         editText = binding!!.editTextPhone
-        button = binding!!.buttonCod
+        button = binding!!.buttonEnterCod
         button.setOnClickListener{onButtonClickListener()}
     }
 
     fun onButtonClickListener() {
 
     }
-
 }
