+++ app/src/main/res/layout/loading_rates.xml
@@ -1,7 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+<androidx.constraintlayout.widget.ConstraintLayout
     android:id="@+id/loading_frame_layoute"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
@@ -50,3 +52,4 @@
     </HorizontalScrollView>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
