+++ app/src/main/res/layout/registr_fragment_welcome.xml
@@ -1,44 +1,45 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <LinearLayout
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
+        android:layout_height="match_parent">
 
-        <TextView
+        <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            style="@style/DescriptionEnterCodeStyle"
-            android:text="@string/welcome" />
+            android:layout_height="match_parent"
+            android:orientation="vertical"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent">
 
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            style="@style/DescriptionTableStyle"
-            android:layout_marginTop="10dp"
-            android:text="@string/for_further_work" />
+            <TextView
+                style="@style/DescriptionEnterCodeStyle"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/welcome" />
 
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            style="@style/DescriptionTableStyle"
-            android:layout_marginTop="20dp"
-            android:text="@string/you_can_fill_in_the_data" />
+            <TextView
+                style="@style/DescriptionTableStyle"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="10dp"
+                android:text="@string/for_further_work" />
 
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            style="@style/DescriptionBodyStyle"
-            android:layout_marginTop="20dp"
-            android:textSize="20sp"
-            android:text="@string/required_documents" />
+            <TextView
+                style="@style/DescriptionTableStyle"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="20dp"
+                android:text="@string/you_can_fill_in_the_data" />
+
+            <TextView
+                style="@style/DescriptionBodyStyle"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="20dp"
+                android:text="@string/required_documents"
+                android:textSize="20sp" />
 
 
             <TableRow
@@ -49,9 +50,10 @@
                 <ImageView
                     android:layout_width="25dp"
                     android:layout_height="25dp"
-                    android:layout_weight="0.09"
                     android:layout_marginLeft="7dp"
+                    android:layout_weight="0.09"
                     app:srcCompat="@mipmap/round_mark" />
+
                 <TextView
                     style="@style/DescriptionTableStyle"
                     android:layout_width="wrap_content"
@@ -69,8 +71,8 @@
                 <ImageView
                     android:layout_width="25dp"
                     android:layout_height="25dp"
-                    android:layout_weight="0.2"
                     android:layout_marginLeft="10dp"
+                    android:layout_weight="0.2"
                     app:srcCompat="@mipmap/round_mark" />
 
                 <TextView
@@ -90,8 +92,8 @@
                 <ImageView
                     android:layout_width="25dp"
                     android:layout_height="25dp"
-                    android:layout_weight="0.3"
                     android:layout_marginLeft="19dp"
+                    android:layout_weight="0.3"
                     app:srcCompat="@mipmap/round_mark" />
 
                 <TextView
@@ -103,5 +105,6 @@
                     android:text="@string/vehicle_reg_certif" />
 
             </TableRow>
-    </LinearLayout>
-</androidx.constraintlayout.widget.ConstraintLayout>
+        </LinearLayout>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
